+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Bump `k8s.io/kubernetes` to `v1.20.7`.
+### Fixed
+- **BREAKING CHANGE**, the Pod metric evaluation now correctly looks at the `averageValue` metric target field rather
+than the incorrect `value` target field.
 
 ## [v0.7.0] - 2021-08-15
 ### Changed