+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v0.8.0] - 2021-12-27
 ### Added
 - Support for `argoproj.io/v1alpha1` `Rollout` resource.
 ### Changed
@@ -51,7 +53,8 @@ than the incorrect `value` target field.
 ### Added
 - Restructured the Horizontal Pod Autoscaler to work within a Custom Pod Autoscaler.
 
-[Unreleased]: https://github.com/jthomperoo/horizontal-pod-autoscaler/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/jthomperoo/horizontal-pod-autoscaler/compare/v0.8.0...HEAD
+[v0.8.0]: https://github.com/jthomperoo/horizontal-pod-autoscaler/compare/v0.7.0...v0.8.0
 [v0.7.0]: https://github.com/jthomperoo/horizontal-pod-autoscaler/compare/v0.6.0...v0.7.0
 [v0.6.0]: https://github.com/jthomperoo/horizontal-pod-autoscaler/compare/v0.5.0...v0.6.0
 [v0.5.0]: https://github.com/jthomperoo/horizontal-pod-autoscaler/compare/v0.4.0...v0.5.0